requirements.txt
aiofiles==23.2.1
aiogram==3.3.0
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.2.0
attrs==23.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
Deprecated==1.2.14
distro==1.9.0
frozenlist==1.4.1
h11==0.14.0
httpcore==1.0.3
httpx==0.26.0
idna==3.6
magic-filter==1.0.12
multidict==6.0.5
netaddr==1.1.0
openai==1.12.0
pydantic==2.5.3
pydantic-settings==2.1.0
pydantic_core==2.14.6
python-dotenv==1.0.1
requests==2.31.0
setuptools==68.2.2
sniffio==1.3.0
tqdm==4.66.2
typing_extensions==4.9.0
urllib3==2.2.0
wheel==0.41.2
wrapt==1.16.0
yarl==1.9.4
yookassa==3.0.1
